+++ src/ts/VAST/Models/VastCompanionAdStaticResource.ts
@@ -1,10 +1,7 @@
 import { Model } from 'Core/Models/Model';
-import { IVastCreativeCompanionAd } from 'VAST/Models/IVastCreativeCompanionAd';
+import { IVastCreativeCompanionAd, VastCompanionAdType } from 'VAST/Models/IVastCreativeCompanionAd';
 
 interface IVastCompanionAdStaticResource extends IVastCreativeCompanionAd {
-    width: number;
-    height: number;
-    type: string;
     staticResourceURL: string | null;
     creativeType: string | null;
     companionClickThroughURLTemplate: string | null;
@@ -14,8 +11,8 @@ interface IVastCompanionAdStaticResource extends IVastCreativeCompanionAd {
 
 export class VastCompanionAdStaticResource extends Model<IVastCompanionAdStaticResource> {
 
-    constructor(id: string | null, height: number | null, width: number | null, creativeType?: string | null, staticResourceURL?: string | null, companionClickThroughURLTemplate?: string | null, companionClickTrackingURLTemplates?: string[], trackingEvents?: { [eventName: string]: string[] }) {
-        super('VastCreativeCompanionAd', {
+    constructor(id: string | null, height: number, width: number, creativeType?: string | null, staticResourceURL?: string | null, companionClickThroughURLTemplate?: string | null, companionClickTrackingURLTemplates?: string[], trackingEvents?: { [eventName: string]: string[] }) {
+        super('VastCompanionAdStaticResource', {
             id: ['string', 'null'],
             width: ['number'],
             height: ['number'],
@@ -30,7 +27,7 @@ export class VastCompanionAdStaticResource extends Model<IVastCompanionAdStaticR
         this.set('id', id || null);
         this.set('width', width || 0);
         this.set('height', height || 0);
-        this.set('type', '');
+        this.set('type', VastCompanionAdType.STATIC);
         this.set('creativeType', creativeType || null);
         this.set('staticResourceURL', staticResourceURL || null);
         this.set('companionClickThroughURLTemplate', companionClickThroughURLTemplate || null);
@@ -67,12 +64,12 @@ export class VastCompanionAdStaticResource extends Model<IVastCompanionAdStaticR
         return this.get('id');
     }
 
-    public getCreativeType(): string | null {
-        return this.get('creativeType');
+    public getType(): VastCompanionAdType {
+        return this.get('type');
     }
 
-    public getType(): string {
-        return this.get('type');
+    public getCreativeType(): string | null {
+        return this.get('creativeType');
     }
 
     public getStaticResourceURL(): string | null {